initial code cleanup, separting out models into their own files
