+++ tests/fixtures/model_fixtures.py
@@ -237,8 +237,8 @@ def models_trained_for_one_step(request,
                                                   item_dense_layers_dims=[15, 10],
                                                   embedding_dropout_p=0.05,
                                                   dense_dropout_p=0.1,
-                                                  lr=1e-1,
-                                                  bias_lr=1e-2,
+                                                  lr=1e-2,
+                                                  bias_lr=1e-1,
                                                   optimizer='adam',
                                                   bias_optimizer='sgd',
                                                   weight_decay=1e-7,
